Structural cleanup of the HDR colorpass work and a reference-white test bug fix
Branch-local: leak-safe Metal color_target check, regrouped/renamed colorpass
state in app.h, fixed a stale shader comment, documented the single-window
assumption in the display color-space cache, and moved supports_hdr() out
from under the colorpass banner comment.

Codebase-wide: split HDRViewApp's ~1500-line constructor into named
setup_*() methods, moved the colorpass GPU-resource lifecycle into a new
app-colorpass.cpp, grouped app.h's member variables by concern, and moved
several non-hot-path colorspace.h functions into colorspace.cpp to reduce
rebuild fanout.

Also: m_render_pass/m_shader are now unique_ptr (matching the colorpass
resources) with proper BeforeExit cleanup instead of being leaked, and
tests/test_colorspace.cpp and colorspaces.sglsl no longer reference the
stale pre-fix 219 nits constant now that HDR_REFERENCE_WHITE_NITS is 203.
